chore(lotto): 개별 상수 파일 하나의 index 파일로 생성
