--- src/constants/lotto.js
@@ -1,8 +1,0 @@
-const LOTTO = {
-  NUMBER_COUNT: 6,
-  MIN_NUMBER: 1,
-  MAX_NUMBER: 45,
-  PRICE: 1000,
-};
-
-module.exports = LOTTO;
